requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
pandas==2.2.2
geopandas==0.14.4
requests==2.31.0
shapely==2.0.4
polyline==2.0.1
pyproj==3.6.1
scipy==1.13.1
networkx==3.3
numpy==1.26.4
python-dotenv==1.0.1
aiofiles==23.2.1